+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 
 * cli: Lazy import dependencies to speed up startup.
 
+## 5.1.5 - 2022-06-05
+
+### Fixed
+
+* config: Fixed crash when rollback hook is about to be called.
+
 ## 5.1.4 - 2022-06-02
 
 ### Fixed